add KeyValueInput component
