+++ ui/src/pages/namespace/Explorer/Page/poc/BlockEditor/QueryProvider/QueryForm.tsx
@@ -2,10 +2,10 @@ import { Controller, useForm } from "react-hook-form";
 import FormErrors, { errorsType } from "~/components/FormErrors";
 import { Query, QueryType } from "../../schema/procedures/query";
 
-import { ArrayForm } from "~/components/Form/Array";
 import { Fieldset } from "~/components/Form/Fieldset";
 import { FormEvent } from "react";
 import Input from "~/design/Input";
+import { KeyValueInput } from "../components/FormElements/QueryParamsForm";
 import { zodResolver } from "@hookform/resolvers/zod";
 
 type QueryFormProps = {
@@ -44,48 +44,13 @@ export const QueryForm = ({
       <Fieldset label="url" htmlFor="url">
         <Input {...register("url")} id="url" />
       </Fieldset>
-      <Fieldset label="Query Parameters">
-        <Controller
-          control={control}
-          name="queryParams"
-          render={({ field }) => (
-            <ArrayForm
-              defaultValue={field.value || []}
-              onChange={field.onChange}
-              emptyItem={{ key: "", value: "" }}
-              itemIsValid={(item) =>
-                !!item && Object.values(item).every((v) => v !== "")
-              }
-              renderItem={({ value: objectValue, setValue, handleKeyDown }) => (
-                <>
-                  {Object.entries(objectValue).map(([key, value]) => {
-                    const typedKey = key as keyof typeof objectValue;
-                    return (
-                      <Input
-                        key={key}
-                        placeholder={
-                          typedKey === "key"
-                            ? "Parameter name"
-                            : "Parameter value"
-                        }
-                        value={value}
-                        onKeyDown={handleKeyDown}
-                        onChange={(e) => {
-                          const newObject = {
-                            ...objectValue,
-                            [key]: e.target.value,
-                          };
-                          setValue(newObject);
-                        }}
-                      />
-                    );
-                  })}
-                </>
-              )}
-            />
-          )}
-        />
-      </Fieldset>
+      <Controller
+        control={control}
+        name="queryParams"
+        render={({ field }) => (
+          <KeyValueInput field={field} label="Query Parameters" />
+        )}
+      />
     </form>
   );
 };
